requirements.txt
langchain
langchain-community
langchain-groq
fastapi
uvicorn
python-dotenv